+++ R/netleague.R
@@ -111,7 +111,10 @@
 #' A.
 #' 
 #' R function \code{\link{netrank}} can be used to change the order of
-#' rows and columns in the league table (see examples).
+#' rows and columns in the league table (see examples). If argument \code{seq}
+#' is a \code{netrank} object, treatments are ordered by decreasing P-scores,
+#' SUCRAs, or probabilities of being best, or by increasing mean or median
+#' ranks.
 #'
 #' @return
 #' An object of class \code{netleague} with corresponding \code{print}
@@ -308,8 +311,16 @@ netleague <- function(x, y,
       ranking.c <- seq$ranking.common
       ranking.r <- seq$ranking.random
       ##
-      seq.c <- setseq(names(ranking.c)[rev(order(ranking.c))], x$seq)
-      seq.r <- setseq(names(ranking.r)[rev(order(ranking.r))], x$seq)
+      method <- if (is.null(seq$method)) "P-score" else seq$method
+      sign <- -1
+      ##
+      if (method %in% c("mean", "median"))
+        sign <- 1
+      ##
+      seq.c <- if (is.null(ranking.c)) x$seq else
+        setseq(names(ranking.c)[order(sign * ranking.c)], x$seq)
+      seq.r <- if (is.null(ranking.r)) x$seq else
+        setseq(names(ranking.r)[order(sign * ranking.r)], x$seq)
     }
     else
       seq.c <- seq.r <- setseq(seq, x$seq)